Create README.md
This project analyzes a telecommunication dataset to understand customer churn. The primary goal is to determine the percentage of customers who have churned and identify the states with the highest churn rates. Additionally, the project explores several factors contributing to customer churn, providing insights that can help improve retention strategies